Use %PRIu64 to print uint64_t instead of %llu.

+++ src/apps/replication/lib/replica_init.cpp
@@ -106,7 +106,7 @@ error_code replica::initialize_on_load(const char* dir, const char* app_type, bo
     {
         // GCed later
         char newPath[256];
-        sprintf(newPath, "%s.%llu.err", dir, static_cast<long long unsigned int>(dsn_now_us()));
+        sprintf(newPath, "%s.%" PRIu64 ".err", dir, dsn_now_us());
         if (dsn::utils::filesystem::rename_path(dir, newPath, true))
         {
             derror("move bad replica from '%s' to '%s'", dir, newPath);
